requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-groq
validators==0.28.1
numexpr
wikipedia